requirements.txt
# python3.9
# pip install torch==2.2.2 torchvision==0.17.2 torchaudio==2.2.2 --index-url https://download.pytorch.org/whl/cu121
diffusers==0.31.0
transformers==4.51.3
xformers==0.0.25.post1 
accelerate==1.10.1
sentencepiece==0.2.0
tokenizers==0.21.4
bitsandbytes==0.47.0
fvcore==0.1.5.post20221221
numba==0.60.0
omegaconf==2.3.0
opencv-python-headless==4.12.0.88 
pytorch-lightning==2.2.0 
qwen-vl-utils==0.0.11
protobuf==6.32.0
wandb==0.21.1
einops
loralib==0.1.2
scipy==1.13.1
pandas==2.3.2
pydantic==2.11.7
pycocotools
timm==1.0.19
openai-clip==1.0.1
gdown